stm:mali: add IRQs resource discovering by name.
In case of the Mali IRQs resources associated to each graphical
processors are not ordered as the driver is expecting,
they will come searched by interrupt name.
This to supported devicetree support.

At probe time, these resources will be searched by using the reg-name
plus the "_IRQ" suffix.
For example, if the reg-name is Mali_GP its interrupt must be
described as: Mali_GP_IRQ.

Signed-off-by: Angelo Castello <angelo.castello@st.com>
Reviewed-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>