stm_nand_bch: tweak handling of erased pages
Due to the BCH ECC format, the NAND Controller will return an uncorrectable ECC
error when reading an erased page.  In order to differentiate between a genuine
uncorrectable ECC error and that generated from reading an erased page, the data
returned by the Controller is tested against the erased page state.

For many newer devices, it is now deemed within specification for an erased page
to include a number of bits at '0', either as a result of read-disturb behaviour
or 'stuck-at-0' failures.  The empty-page test must therefore tolerate a certain
number of bits at '0'.  Details on exactly what this number should be are rather
vague.  For the moment, we have settled on a relatively conservative value
matching the number of BCH sectors per page (i.e. 2 for a LP device, 4 for a VLP
device, and 8 for a VVLP device).  Note, if the BCH ECC format had been
consistent on an erased page, bits at '0' would have been automatically
corrected up to the ECC strength (i.e. 18 or 30).

This patch updates the code such that bits at '0' are "corrected" for data that
passes the erased-page test.  This approach is closer to how other MTD drivers
operate, where the ECC scheme would correct bits at '0' in an erased page.  For
genuine uncorrectable ECC errors, the original uncorrected page data is
returned, in accordance with the MTD API.

Signed-off-by: Angus Clark <angus.clark@st.com>