stm_nand_bch: add bit-flips threshold for '-EUCLEAN' condition
This patch adds support for a bit-flips threshold which is used to determine the
point at which mtd_read() should return '-EUCLEAN'.  The higher layers
(e.g. UBI) use the '-EUCLEAN' return code as an indication that the data may be
deteriorating, triggering various counter-measures such as relocating the data,
and torture testing the block to see if it should be retired.

Previously, '-EUCLEAN' would be returned when one or more bit-flips had been
corrected.  While appropriate for 1-bit ECC NAND, this approach can lead to
unnecessary wear on multi-bit ECC NAND and compromise the effective life-span of
the device.  The bit-flips threshold helps prevent this situation by only
returning '-EUCLEAN' when deemed necessary.

A similar strategy has been adopted recently on linux mainline:
edbc454 "mtd: driver _read() returns max_bitflips; mtd_read() returns -EUCLEAN"
and
d062d4e "mtd: bitflip_threshold added to mtd_info and sysfs"

This patch represents a minimal implementation, applied uniquely to the
stm-nand-bch driver.  The bit-flips threshold is specified via the platform
data, with an equivalent device tree binding.  The value should be in the range
1 to <ecc-strength> where <ecc-strength> is 18 or 30, depending on the BCH ECC
mode in operation.  A value of 0 is interpreted by the driver as <ecc-strength>.

Although the infrastructure has been put in place for tuning the 'EUCLEAN'
condition, there is little information on how best to determine the threshold.
By treating a value of 0 as <ecc_strength>, we have implicitly adopted the same
default as that used in linux mainline.  While generally accepted to be an
improvement over the previous behaviour (essentially bitflip_threshold = 1), it
is likely that the optimum value will depend on a number of factors including
the ECC specification of the NAND device, the strength of the ECC scheme in
operation, and the target application.  In addition, various trade-offs will
need to be considered, such as longevity versus data integrity.  This remains an
open point and will require input from the NAND Vendors and the MTD community.

Signed-off-by: Angus Clark <angus.clark@st.com>