stm: stmmac: use bugged_sg instead of disabling full tx coe
for alicante cut 1 during the bringup stage
it has has been noticed that the hw csum had some problems
so a new platform field was added to disable this capability
(although supported by the HW cap register).

Now this seems to be the right fix because as I suspected
and thx to Aditi now confirmed, the HW has only problems
to do the csum on segments. So we can enable the HW csum
but disable scatter-gather. To do that the bugged_tx_coe
has been substituted by bugged_sg and the driver has been
changed to enable/disable the SG w/o turing-off the whole
csum.

Reported-by: Aditi SHARMA <Aditi-hed.SHARMA@st.com>
Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>