stm: flashss: extend the flashSS adding mmc core support
This patch adds the support to manage the Arasan mmc host controller
from within the FlashSS IP.
Infact, using the FlashSS we can program the Arasan Host controller resisters.
This is mandatory on some platforms like Monaco to allow the SDHCI to work with
a modern controller otherwise it will find a basic configuration suitable for
booting from eMMC. The new subset of registers starts from the Arasan HC base
address + <offset> (0x400).

Note that there also are some dependencies between the mmc core and the FlasSS,
for example, programming the DLL (the register is in the Top FlashSS memory
space) the stack has to check if the procedure completed fine looking at another
register in the MMC CORE memory space.

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>