stm:telss Prevent fifo errors immediately on TDM start
The TDM player must always be on in order to provide a clock to the TELSS
hardware. When no data is being transferred this means the TDM will be in
an under run state.

When we start the TDM player we may immediately get an under run as soon as
the fifo error interrupt is enabled. This is because the FDMA has not yet
copied any data to the fifo but should be in the process of doing so. In
this scenario we can safely ignore the fifo error.

This patch counts how many fifo errors have occurred and only processes an
actual fifo error when the number exceeds a defined threshold.

This patch also adds a missing macro for accessing the number of samples
present in the FIFO.

Signed-off-by: John Boddie <john.boddie@st.com>