stm:fdma:telss Improve latency starting a second handset call
The purpose of this patch is to allow the TELSS driver to start a call on
a particular handset faster. Previously if a call was already active for
another handset, then when starting the second call we would have to wait
for the FDMA to start processing node 0 of the transfer before the second
call became active. This patch allows the second call to become active on
the very next node of the transfer, significantly reducing latency.

To achieve this a number of changes have had to be made:

- Addition of an 'extension' member to struct stm_fdma_desc to store FDMA
extension private data. In the TELSS case we encode the node index and the
total number of nodes in the FDMA transfer.

- Addition of 'period_offset' and 'period_stride' to handset config data to
aid in calculating the handset buffer offsets for each transfer node. This
resulted in dropping the 'period_stride' parameter from the function
dma_telss_prep_dma_cyclic().

- Checks to ensure handset configuration is performed before attempting to
prepare a TELSS transfer.

Signed-off-by: John Boddie <john.boddie@st.com>