stm:telss Improve call starting latency
This patch implements the TELSS driver modifications resulting from the
FDMA driver changes in order to allow handset calls to start faster.

The aim is that when a handset call is already active, another handset call
can be started within the specified duration of an audio period (currently
this is 10ms). This can happen as when starting the second call, we
re-program the handset buffers to start at the very next audio period
rather than waiting for the FDMA transfer to wrap around to the first audio
period.

Signed-off-by: John Boddie <john.boddie@st.com>