stm:telss Implement a cached dma position
This patch changes the way the TELSS driver calculates the dma position,
used for updating the ALSA buffer offsets. Previously, an approximate
dma offset was calculated for each handset during the dma callback, even
although each handset would essentially be at the same dma offset.

What happens now is that the dma callback calculates an approximate dma
offset once that will be valid for each handset. This offset is cached,
and when ALSA asks the TELSS driver frame offset, this cached value is
returned.

The benefits are that we reduce the time within the FDMA driver where
interrupts may be masked, and also some CPU usage.

Signed-off-by: John Boddie <john.boddie@st.com>