stm:telss Do not stop dma on xrun to maintain data continuity
This patch relates to Bugzilla 29839.

The purpose of this patch is, in the case of xrun, to defer stopping the
dma until the pcm device is closed. Whilst the driver is in xrun, for
playback, null samples will be sent to the uniperipheral in multiples of
the audio period size and for capture the data will be dropped in multiples
of the audio period size.

The benefit of this is that on xrun, data continuity can be maintained
between the Lantiq DuSLIC and the SFT software stack. By data continuity
I mean that all data transferred to or from the uniperipheral will remain
in multiples of the audio period and the xrun duration will always be a
multiple of the audio period.

On cases where we detect that playback has been stopped normally, i.e. the
driver stop function is called when the driver is in the draining state,
we stop the dma normally.

Normally stopping capture is always treated as an xrun. This is because
as there is always data to capture it never enters a draining state.

Signed-off-by: John Boddie <john.boddie@st.com>