stm:telss Turn uniperipheral reader off when not in use
When there are no valid calls, the uniperipheral reader should be turned
off to conserve power, and when there is a valid call, the uniperipheral
reader should be turned back on.

The uniperipheral player should never be turned off as it supplies the
clock to both the Lantiq device and the uniperipheral reader.

Signed-off-by: John Boddie <john.boddie@st.com>