stm:telss Specify number of audio periods in platform configuration
This patch will allow the number of audio periods to be specified in the
platform configuration data instead of being hard coded into the TELSS
driver.

If the specified number of periods cannot be honoured, a message will be
output on driver load indicating how many periods are available.

A minimum of two periods is enforced.

Signed-off-by: John Boddie <john.boddie@st.com>