stm: add docsis queue driver
This is a small driver used for configuring, on some ST platforms, the
Upstream and Downstream queues before using the Virtual Interface
with ISVE driver.

Currently it only enables the Downstream queues but it can be extended
to support further settings (reset etc ... now done by docsis code
running on other CPUs).

It has been necessary to move some logic in the ARM9 side as reported
in the bug 31303 to, in summary, allow a sync among all eSAFEs and
eCMMaster.

It is a platform driver and supports DT model too.
This driver can be used to expose the queue registers via debugFS support:
/sys/kernel/debug/docsis_queue

Note: it has been avoided to do these settings inside ISVE that remains
a virtual Ethernet driver able to only manage the queue 0 ... N, DMA
and relative registers at offset 0x400.

In fact, the task made by docsis_queue driver couldn't be moved in
ISVE where is was not possible decide which queues had to manage
the Downstream/Upstream registers that are share among all the
queues.

    V2: fixed the module when build as dym module (thx to
        Arif Siddiqui <arif.siddiqui@st.com>

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>