stm: fix mmc on b2116
The commit 82f08c6ff958063c93be enhances the MMC for DT adding the way to pass
removable card field. This patch fixes the build of the b2116 and b2105 that
fails becasue the mmc_platform_data has been removed and also is to pass the
st,mmc-non-removable for example for the eMMC_1 available in the b2116.

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>