stm: stid127: fix the Uart2 RX pin pad
This patch fixes the gpio pin for the UART2_RXD signal:
STiD127 pad configuration file represents UART2 RX pin as
PIO17.5, it should be PIO20.2.
Strangely (perhaps because the pin is configured by default as input)
the serial, which is the console on all Alicante boards,
still works even without this patch.

Signed-off-by: Nunzio Raciti <nunzio.raciti@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>