stm: stih416: change the RTC clock rate from 300KHz to 400KHz
It is observed that clk_set_rate from rtc-stm driver fails for 300KHz
value, because when the VCO (parent clock) is running @480 MHz, it
cannot generate a frequency of 300KHz for its child.
This clock as a constraint to be >= 300KHz as shared with the thermal
sensor, so setting it at 400KHz is fine and within RTC specification.

[carmelo.amoroso@st.com: improved description with comments from
bugzilla]
Signed-off-by: Shiraz Hashim <shiraz.hashim@st.com>
Signed-off-by: Carmelo Amoroso <carmelo.amoroso@st.com>