stm: sysconf: Make debugfs output safe against STBus errors
Increasingly we are seeing problems where the sysconf registers return
STBus errors when a register which is not implemented is read/written.
Using the generic abort handling which was added in a previous commit
it is now possible to catch these errors and report the register as
RESERVED, instead of raising an exception.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>
Signed-off-by: Carmelo Amoroso <carmelo.amoroso@st.com>