stm: pio-control: Fix debug support when empty control fields
This is a follow up to commit a01478d99233c04a64d33a771b78d256a6b8fac9
    stm: pio-control: allow empty control fields
which fixes the debugfs output in the case of missing fields.

Note that the default values (used when a sysconf is missing) are only
correct for the FlashSS on Cannes. This will need fixing if we see
similar but different changes in the future.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>
Signed-off-by: Carmelo Amoroso <carmelo.amoroso@st.com>