stm: pio-control: Allow user to force delays for bidir pins
Currently the code doesn't allow the user to configure delays on
bidirectional pins. While I'm still not convinced it makes sense,
this patch allows the user to experiment with delays to work
around board issues.

In non-DT case two extra fields have been added to the existing
struct stm_pio_control_retime_config:
	unsigned int force_delay:1;
which must be set, and:
	unsigned int force_delay_innotout:1;
which is needed for PIO banks with the "dedicated" style of retiming
control registers - this is ignored for the "packed" style.

If configuring this from the device tree and extra field has been added
to the retiming attributes, for example:
   <&PIO19 3	BIDIR ALT5 BYPASS CLK_B FORCE_OUTPUT_DELAY 1000>
similarly FORCE_INPUT_DELAY.

[carmelo.amoroso@st.com: fixed checkpatch warnings due to 80 columns]
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>
Signed-off-by: Carmelo Amoroso <carmelo.amoroso@st.com>