stm: spi_stm: remove support for custom chip select handlers
The spi_stm driver included support for custom chip select handlers which could
be passed in as callbacks via the platform data.  Originally, this was intended
to be used in cases where the CS lines were driven by something other than
standard PIOs (e.g. an EPLD).  However, when the ST PIO support was migrated to
the GPIO framework, there was no longer any need for custom chip select
handlers, since equivalent functionality could be mapped onto virtual GPIOs.

This patch removes support for custom chip select handlers -- it is no longer
required, it paves the way for DT support, and greatly simplifies the code (in
many cases, the SSC configuration structure can be removed altogether).

Note, the only existing user of the '.spi_chipselect' support is the b2078.  The
use here is not entirely obvious, since it appears to install a single GPIO CS
to be used for all device on the SPI bus.  Clearly this is not going to work on
buses with more than one device.  The most likely explanation is that it was a
quick hack to get something working.  We remove the custom chip selects here,
with a view to add correct support as and when the requirement is made clear.

Signed-off-by: Angus Clark <angus.clark@st.com>
Signed-off-by: Carmelo Amoroso <carmelo.amoroso@st.com>