stm: spi_stm: use spi_board_info->controller_data to specify the chip select
This patch changes which 'spi_board_info' field is used to specify the chip
select from '.chip_select' to '.controller_data'.  This is in-line with other
spi drivers, which means the same configuration data can now be used for both
spi_stm and spi-gpio drivers.  To ensure full compatibilty, the spi_stm driver
now supports SPI_GPIO_NO_CHIPSELECT, as well as STM_GPIO_INVALID, to denote "No
Chip Select" is used.

In the base stlinux distribution, the b2044 is the only board that registers an
SPI device.  The device is specified with ".chip_select = 0', which actually
corresponds to PIO0[0].  This is almost certainly not what the author originally
intended, but for the purpose of this patch we keep the same value when porting
to ".controller_data = 0".  (An update to what we believe the chip select should
be will be made in a subsequent patch!)

Signed-off-by: Angus Clark <angus.clark@st.com>
Signed-off-by: Carmelo Amoroso <carmelo.amoroso@st.com>