stm: b2044: update "spicore" chip select
The chip select for the SPI device "spicore" is specified as ".controller_data =
0" which coresponds to PIO0[0].  However, it is more likely that the origianl
intention was to specify "No Chip Select", and therefore STM_GPIO_INVALID should
have be used (or SPI_GPIO_NO_CHIPSELECT, which is now the preferred option) .
However, on reviewing the board's schematic, it appears that PIO16[3] is the
correct chip select for "spicore".

Unfortunately, this is not easy for us to test since "spicore" is not part of
the stlinux distribution (and not yet in SDK2 either).

Signed-off-by: Angus Clark <angus.clark@st.com>
Signed-off-by: Carmelo Amoroso <carmelo.amoroso@st.com>