isve: request fewer upstream interrupts and adjust free threshold
The driver was requesting some interrupts which it did not handle in the
ISR. Some interrupts reasons where not needed, so those are masked out now.

UPIIM_INQ_CTL_FILL_FIFO_EMPTY was masked out before this fix as well, it
was the high bit in the '0x3D' value of the FILL_EMPTY threshold. The bit
has been removed from there and the interrupt is now masked explicitly.

The threshold of the freed FIFO is lowered to 16 packets. The existing
threshold of 29 was very close to the FIFO size of 32. If the freed FIFO is
every completely full the UPIIM will stop processing packets from the filled
FIFO until room becomes available. Lowering the freed threshold will reduce the
chance of this happening.

Signed-off-by: Martin Habets <martin.habets@st.com>
Acked-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>