isve: Only call isve_tx() if there is something in the freed FIFO
upiim_interrupt() was always returning "in_packet", even if the freed FIFO
was empty. This caused too many calls to isve_tx(), which would wake up the
interface even when it should not be.

upiim_interrupt() now also returns "no_packet" if no work is needed, and
"in_packet_dropped" when the filled FIFO is full. A later patch will add the
ISR handling for "in_packet_dropped".

Signed-off-by: Martin Habets <martin.habets@st.com>
Acked-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>