isve: Empty the freed FIFO based on the hardware status
The existing ISR loop using cur_tx was not SMP safe, as cur_tx can be changed
on the other core. When this happened the UPIIM_INQ_INT_STAT_FREE_EMPTY_FIFO
interrupt would never get resolved, and the kernel would run the ISR forever.

In this patch we remove the use of cur_tx in the ISR, and in stead empty the
freed FIFO based on the hardware status register.
upiim_freed_tx_add() now returns 0 if the freed FIFO is empty, and it ensures
the status register is always up-to-date.

upiim_fill_tx_add() also ensures the status register is always up-to-date.
Without this we can delay a UPIIM_INQ_INT_STAT_FILL_FULL_FIFO interrupt until
another upiim interrupt comes along (which could possibly never happen).

Signed-off-by: Martin Habets <martin.habets@st.com>
Acked-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>