isve: Minor preemption fixes
In an SMP system dirty_tx and cur_tx can be off by 1. This can result
in a negative return value from isve_tx_avail(), with unexpected results.
The same can happen when cur_tx overflows it's maximum value.
Since all inputs to the calculation are unsigned, so to solve this we return
an unsigned value as well.

In case of a TX timeout we should grab the spinlock as well, to prevent for
example isve_xmit from running.

Signed-off-by: Martin Habets <martin.habets@st.com>
Acked-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>