stm_of: Add clk disable support and error handling
This patch add new feature to disable the clock via stm_of sequences, By
default a sequence of type==clock configures the clock and enables it.

The reason for this new binding is:
 - power down sequences might be interested in disabling the clock.
 - some drivers want to just configure parenting of the device clock and
   manage the clock in the driver code.

To disable a clock you need a sequence like below in device tree:

step1 {
	clk-name = "clk-xyz";
	clk-disable;
	type = "clock";
};

This patch also adds error checks which was missing in the original
patches.

Buglink: https://bugzilla.stlinux.com/show_bug.cgi?id=35675

Signed-off-by: Srinivas Kandagatla <srinivas.kandagatla@st.com>
Signed-off-by: Carmelo Amoroso <carmelo.amoroso@st.com>