stm_gpio: Rework stm gpio interface for 3.4 kernel
This patch tries to resurrect the stm_gpio interface for the 3.4 kernel.
The way that interrupts are managed has changed in latter kernels. This
patch updates the calls to the later equivalents.

However, it is not totally compatible, the global IRQ_DISABLED flag has
gone. I've introduced an equivalent STPIO_IRQ_DISABLED flag which will
have the same semantics. It should be trivial to change the caller of
the function to use this new flag. I have chosen to do this rather than
introduce a IRQ_DISABLED flag in case this old flag is used elsewhere.

I've looked at the STAPI usage of this function, and it also uses the
old WAKE flag, but this cannot have worked for some time as it was never
passed through to the lower levels as far as I can see.

Signed-off-by: David McKay <david.mckay@st.com>