stm/mpe41: avoid duplicate resource init that caused a oops
This fix aligns mpe41.c with mpe42.c file and only do machine init
of resources in case device tree is not used, based on CONFIG_OF flag.

This was required in case of PMU initialisation that failed and caused
a oops when CONFIG_HW_PERF_EVENTS was enabled on stih415.

Signed-off-by: Laurent Meunier <laurent.meunier@st.com>
Signed-off-by: Carmelo Amoroso <carmelo.amoroso@st.com>