stm: isve: enforce the tx fault recovery function
in case of hard error the tx ISR calls the same
function used when the tx WDT expires. This patch
improves this function. In fact, tx resources are
freed and the tx is restarted. This scenario remains
untested on my side because no way to reproduce it.
This in fact should be a really uncommon case.

Cc: Martin Habets <martin.habets@st.com>
Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>