stm: b2116: fix miphy-0 configuration as PCIe
On b2116 there is not SATA port, and the only MiPHY available is
connected to the PCIe controller, so current board configuration is
wrong.
This patches fixes it by configuring MiPHY-0 in PCIe mode and adding the
the call to configure PCIe instead of SATA.
It requires to have the '.tx_pol_inv' field set.

Buglink: https://bugzilla.stlinux.com/show_bug.cgi?id=36195

Tested-by: Lionel Debieve <lionel.debieve@st.com>
Tested-by: Nunzio Raciti <nunzio.raciti@st.com>
CC: Bossen Wu < bossen.wu@st.com>
Signed-off-by: Carmelo Amoroso <carmelo.amoroso@st.com>