stm_pcie: Improve allocation of MSI irq numbers
Previously, the irq range allocated to the MSI irqs was in effect
allocated statically. This was a pain as you had to know in advance
where the empty spaces in the irq ranges were and pass them into the
driver.

Fortunately, we can now allocate the irqs using irq_alloc_descs(), which
will search the irq range looking for a hole of a suitable size. This
removes the need to know where the irqs are going to be in advance.

We no longer need to check if the handler is in use, and the check for
having zero endpoints is redundant as it cannot happen since we have to
have at least one irq allocated. The irqs in the stucture are changed to
int from unsigned as negative values are now used in the code, and it
makes it consistent as well.

We start searching for free irqs from the mux irq, as if we start from
zero on the ARM we will allocate some of the MSI irqs below 32. This
could potentially confuse people, as irqs < 32 are traditionally
"special".  It would also mean that some irqs would end up at the bottom
and some at the top if there is more than one PCIe bus present.

Signed-off-by: David McKay <david.mckay@st.com>