net: stmmac: fix and better tune the default buffer sizes
This patch is to fix and tune the default buffer sizes.
It reduces the default bufsize used by the driver from
2048 to 1518 (taking into account the extra 4 bytes in case of VLAN).

Patch has been tested on both ARM and SH4 platform based.

Also it has been noticed another point.
The netdev_alloc_skb reserves 16 bytes, and typically NET_IP_ALIGN means
we reserve 2 more, this pushes us to allocate from the next larger slab
size because the default bufsize is set to 2048 bytes.
On some platforms (i.e. STi 7108/09 ...) this is forcing to to increment
the min_free_kbytes also when we work with the std Ethernet MTU.

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>